students/y2336/practical_works/Bondarev_Nikita/Final Work/requirements.txt
Django~=3.1.2
requests~=2.24.0
Pillow~=7.2.0
beautifulsoup4~=4.9.2